requirements.txt
# Backend Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# AI Services
openai>=1.12.0

# Environment
python-dotenv>=1.0.0
